+++ src/main.rs
@@ -1,18 +1,12 @@
-use ansi_term::Style;
-use crossterm::style;
-use serde::__private::de::FlatInternallyTaggedAccess;
 use web_parser::{
     process_string::bond,
-    process_string::bond::{markup_entry, parse_style_sheet, styles_hash, remove_tabs},
     static_data::structs::{ReqPair, StyleChild, StyleMain, TermmlMain},
     webrequest::request::{fetch, get_filename},
 };
 
-use renderer::{debug::ren_debug::DebugRenderer, entry::ren_entry::MainNavigator};
 
 use hard_xml::{XmlRead, XmlWrite};
-use std::{alloc, collections::HashMap, process::Child};
-use ureq::{Response, Transport};
+use std::{alloc, collections::HashMap};
 
 //tracking memory usage
 use cap::Cap;
@@ -27,7 +21,7 @@ fn start() {
     let mut files: HashMap<String, String> = HashMap::new();
     let server_url = String::from("http://127.0.0.1:5500/");
     let url = format!("{}{}", server_url, "test.termml");
-    let filename = get_filename(&url);
+    let _filename = get_filename(&url);
     let mut termss_vec: Vec<String> = vec![];
     let fetched = match fetch(&url) {
         Ok(r) => r,
@@ -112,94 +106,94 @@ fn start() {
 }
 
 //TEMP
-fn construct_termml_vec(
-    markup: TermmlMain, stylesmap: HashMap<String, StyleChild>
-    ) -> Vec<(String, StyleChild)>{
-    let mut vec: Vec<(String, StyleChild)> = vec![];
-    let head_divs = markup.head.value;
-    let body_divs = markup.body.value;
-    match head_divs.class {
-        Some(class) => {
-            let k: String = class.into();
-            let c = stylesmap.get(&k);
-            let style = c.cloned();
-            drop(c);
-            match style {
-                Some(child) => {
-                    println!("pushing vec");
-                    vec.push((head_divs.value.into(), child));
-                },
-                None => {
-                    println!("no pushing vec");
-                    vec.push((head_divs.value.into(), 
-                        StyleChild {
-                            class: String::from("null"),
-                            background: None,
-                            foreground: None,
-                            underline: None,
-                            bold: None,
-                            header: Some(true)
-                        }
-                    ));
-                }
-            }
-        },
-        None => {
-            vec.push((head_divs.value.into(),
-                StyleChild {
-                    class: String::from("null"),
-                    background: None,
-                    foreground: None,
-                    underline: None,
-                    bold: None,
-                    header: Some(true)
-                }
-            ))
-        }
-    }
-    for mut i in body_divs {
-        println!("in loop");
-        i.value = remove_tabs(i.value.clone().to_string()).into();
-        match i.class {
-            Some(class) => {
-                let k: String = class.into();
-                let c = stylesmap.get(&k);
-                let style = c.cloned();
-                drop(c);
-                match style {
-                    Some(child) => {
-                        vec.push((i.value.into(), child));
-                    },
-                    None => {
-                        vec.push((i.value.into(), 
-                            StyleChild {
-                                class: String::from("null"),
-                                background: None,
-                                foreground: None,
-                                underline: None,
-                                bold: None,
-                                header: Some(false)
-                            }
-                        ));
-                    }
-                }
-            },
-            None => {
-                vec.push((i.value.into(),
-                StyleChild {
-                    class: String::from("null"),
-                    background: None,
-                    foreground: None,
-                    underline: None,
-                    bold: None,
-                    header: Some(false)
-                }
-            ))
-            }
-        }
-    }
-    return vec;
-}
+// fn construct_termml_vec(
+//     markup: TermmlMain, stylesmap: HashMap<String, StyleChild>
+//     ) -> Vec<(String, StyleChild)>{
+//     let mut vec: Vec<(String, StyleChild)> = vec![];
+//     let head_divs = markup.head.value;
+//     let body_divs = markup.body.value;
+//     match head_divs.class {
+//         Some(class) => {
+//             let k: String = class.into();
+//             let c = stylesmap.get(&k);
+//             let style = c.cloned();
+//             drop(c);
+//             match style {
+//                 Some(child) => {
+//                     println!("pushing vec");
+//                     vec.push((head_divs.value.into(), child));
+//                 },
+//                 None => {
+//                     println!("no pushing vec");
+//                     vec.push((head_divs.value.into(), 
+//                         StyleChild {
+//                             class: String::from("null"),
+//                             background: None,
+//                             foreground: None,
+//                             underline: None,
+//                             bold: None,
+//                             header: Some(true)
+//                         }
+//                     ));
+//                 }
+//             }
+//         },
+//         None => {
+//             vec.push((head_divs.value.into(),
+//                 StyleChild {
+//                     class: String::from("null"),
+//                     background: None,
+//                     foreground: None,
+//                     underline: None,
+//                     bold: None,
+//                     header: Some(true)
+//                 }
+//             ))
+//         }
+//     }
+//     for mut i in body_divs {
+//         println!("in loop");
+//         i.value = remove_tabs(i.value.clone().to_string()).into();
+//         match i.class {
+//             Some(class) => {
+//                 let k: String = class.into();
+//                 let c = stylesmap.get(&k);
+//                 let style = c.cloned();
+//                 drop(c);
+//                 match style {
+//                     Some(child) => {
+//                         vec.push((i.value.into(), child));
+//                     },
+//                     None => {
+//                         vec.push((i.value.into(), 
+//                             StyleChild {
+//                                 class: String::from("null"),
+//                                 background: None,
+//                                 foreground: None,
+//                                 underline: None,
+//                                 bold: None,
+//                                 header: Some(false)
+//                             }
+//                         ));
+//                     }
+//                 }
+//             },
+//             None => {
+//                 vec.push((i.value.into(),
+//                 StyleChild {
+//                     class: String::from("null"),
+//                     background: None,
+//                     foreground: None,
+//                     underline: None,
+//                     bold: None,
+//                     header: Some(false)
+//                 }
+//             ))
+//             }
+//         }
+//     }
+//     return vec;
+// }
 
 fn _alloced<T: std::fmt::Display>(header: T) -> () {
     //only for development : not suppossed to be in actual build AT ALL
